+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 (Please put an entry here in each PR)
 
 - Show secret keys and tokens as `"**********"` in `Debug` output
+- Ensure list of signed headers is correct when not all headers are signed
 - Use ```$AWS_PROFILE``` to obtain default profile name 
 - Implement `Default` for `Region`
 - Derive Clone for remaining types (affects CloudFront, Route 53 and S3)